Deduplicate executable bit command logic
